Move argspec functions into arg mod. Gate inspect call.
